% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.r
\name{slurm_write_r}
\alias{slurm_write_r}
\title{slurm_write_r}
\usage{
slurm_write_r(id, dir = ".", r_command)
}
\arguments{
\item{id}{job id}

\item{dir}{directory for saving the output R script}

\item{r_command}{R command in string}
}
\description{
Write the accompany R script fot the Slurm sript
}
