% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.r
\name{slurm_write_script}
\alias{slurm_write_script}
\title{slurm_write_script}
\usage{
slurm_write_script(
  id = NA,
  dir = ".",
  partition = "small",
  conda = "r",
  nodes = 1,
  tasks = 1,
  wd = ".",
  walltime = "24:00:00",
  type = "r",
  gpu = 2,
  command = NULL
)
}
\arguments{
\item{id}{job id}

\item{dir}{directory for saving the output R script}

\item{partition}{Slurm partition}

\item{conda}{Conda environment}

\item{nodes}{Number of required nodes}

\item{tasks}{Number of required tasks}

\item{wd}{the working directory}

\item{walltime}{Walltime for the job}

\item{type}{Script type}

\item{gpu}{Number of required GPUs}

\item{command}{Command (only for bash jobs)}
}
\description{
Write Slurm script
}
